backend/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
PyJWT
cloudinary
django-cloudinary-storage[video]
pillow
environs